+++ src/js/renderers/flv.js
@@ -200,8 +200,7 @@ const FlvNativeRenderer = {
 					}
 
 					node.addEventListener(eventName, (e) => {
-						const event = document.createEvent('HTMLEvents');
-						event.initEvent(e.type, e.bubbles, e.cancelable);
+						const event = createEvent(e.type, mediaElement);
 						mediaElement.dispatchEvent(event);
 					});