+++ src/js/renderers/hls.js
@@ -216,8 +216,7 @@ const HlsNativeRenderer = {
 
 					node.addEventListener(eventName, (e) => {
 						// copy event
-						const event = document.createEvent('HTMLEvents');
-						event.initEvent(e.type, e.bubbles, e.cancelable);
+						const event = createEvent(e.type, mediaElement);
 						mediaElement.dispatchEvent(event);
 					});